eve-chromebook: Add support for ROOTLESS_X and linux-yocto
Add support to starting X as a non-root user (why this has to be
enabled through MACHINE.conf is beyond me).

Enable building linux-yocto and set it as default (vs linux-intel).

Fix the way we were disabling gobject data instrospection, should
be wired in by disabling qemu-usermode in MACHINE_FEATURES.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <aehs29@gmail.com>