chromebook-console-keymap: Set the default console keymap as the pixelbook keymap
The kbd package contains a patch that adds a pixelbook keymap,
we want to set this as default when the machine is booting up
automatically.

Load the previously mentioned keymap at runlevels 1, 3 and 5
and also set the font to a larger one for normal humans above
the age of 20 to be able to see whats on the console (coming
from terminus-font-consolefonts).

Signed-off-by: Alejandro Enedino Hernandez Samaniego <aehs29@gmail.com>