xkeyboard-config: Add pixelbook-eve keyboard model
The Pixelbook has a pretty specific keyboard, along with some
key bindings that work with Chrome/Chromium OS.

This patch adds a pixelbook-eve keyboard model to xkeyboard-config
which aims to emulate these key bindings to a certain extent.

It uses the Search key as modifier.

Allows the user to use to upper keys as F1-F11 as in a normal keyboard,
but if the modifier is pressed it sets functions similar to the ones
present on Chrome.
e,g. The first key (after Esc) pressed along with the search key would
trigger the Back event (just like its symbol says), the next key, pressed
along with the search key would trigger a Reload event, due to the lack
of a better choice the next key would open a terminal, brightness and
volume keys would trigger the according events and the last key would
trigger the XF86Tools event.

Also, Backspace pressed with the modifier now becomes Delete, just like
in Chrome.

And finally the arrow keys with the modifier behave the same as in Chrome
as well, Left=Home, Right=End, Up=PgUp and Down=PgDown.

This keyboard model can be set through X either by a config or through
Xs keyboard managing tools.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <aehs29@gmail.com>