xfwm4-config-rm: Fix bug when starting the X server after the first boot
The X server crashes and is unable to start after the first boot
due to either the existence of a file or on of the configs in it.

It causes an assertion error with the i915 driver.

Havent had time to debug which config causes this, so for now,
remove the config file (which is created on every boot anyway),
at different runlevels to allow the X server to start again.

Its worth noting that this apparently only happens when the root
user tries to startx the X server, ROOTLESS_X seems to be working
properly

Signed-off-by: Alejandro Enedino Hernandez Samaniego <aehs29@gmail.com>