tune-skylake: Disable AVX extensions from the Intel ISA
During a build, some of the components such as Gobject Data
Introspection or Chromium, require to run a QEMU to run code
as if it were running on the target.

When using -march=skylake and -mtune=skylake AVX instructions
get enabled by default for optimization, but at this point
QEMU (cpu Skylake-Client) does not support such instructions,
which causes QEMU to crash due to illegal instructions.

Remove the required avx optimizations from the skylak tune
to allow these components to cross-compile correctly.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <aehs29@gmail.com>