pixelbook-image-xfce: NO CHROMIUM, NO PARTY
This layer is meant to support building customized Linux
distributions for Chromebooks (at this point Pixelbook).

It really wouldnt be a very good chromebook layer if it
did not support the Chromium web browser, so enable
chromium-x11 on the XFCE image.

This was not as simple as adding chromium-x11 to the image,
due to the way chromium is cross compiled, the v8 engine
requires to run code on an architecture same as the target,
this is achieved / wired in through running QEMU , but
QEMU does not support the AVX instruction extensions and
crashes while running the mentioned code complaining about
an illegal instruction.

Now that the AVX optimizations have been disabled on the
Skylake tune, Chromium is able to build properly.

Trying to disable the AVX optimizations specifically
just for Chromium also, does not work properly, since
Chromium makes calls to libraries like the C library or
libc++api, which try to run AVX instructions and fail.

Another appoach was to use multiconfig, on which another
multiconfig would build a specific toolchain just for
Chromium, that would generate either the required artifacts
or even an rpm package, that would work but it is not
very efficient.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <aehs29@gmail.com>