mrchromebox: Add recipe to provide MrChromeBoxs scripts to handle firmware on Chromebooks
MrChromeBox provides a set of scripts that allow users to
manipulate the firmware on Chromebook devices, in his own
words the scripts provide the following functionality:

- Install/Update the RW_LEGACY firmware with a newer/working/customized
  version of the SeaBIOS firmware payload
- Install/Update the BOOT_STUB firmware to enable booting directly
  Legacy Boot (using the same payload as RW_LEGACY)
- Install/Update a full custom firmware image (Full ROM),
  built from the latest sources
- Set the Boot Options (GBB Flags) (only for stock ChromeOS firmware)
- Set the devices Hardware ID (only for stock ChromeOS firmware)
- Remove the ChromeOS Developer/Recovery mode Bitmaps
  (only for stock ChromeOS firmware)
- Restore the ChromeOS Bitmaps (only for stock ChromeOS firmware)
- Restore the Stock BOOT_STUB
- Restore the Stock Firmware

The recipe provides all the necessary scripts, although due to
the way the script works it forces an update per exection, this part
has been disabled.

The scripts also try to download necessary utilities (as binaries)
to update/install the firmware, e.g. flashrom, crossystem, cbfstool,
gbb_utility.

There are recipes for each of these tools and all of them are also
built from source and installed on the image, but the script will
still try to use its own versions, at this point it is left to the
user to decide which tool should be used, if the user decides to
use the built from source tools, the scripts have to be manually
modified though, at least for now.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <aehs29@gmail.com>