eve-chromebook: Clean up machine conf after upstreaming skylake changes
The skylake tune changes were sent and accepted upstream, they got merged on:
meta-intel: 0bc64c2ad

With this in mind, base the eve-chromebook machine conf on the intel-skylake-64 machine,
and try to keep the conf as close sas possible to upstream.

We can also now delete the tune includes from this layer, and also a workaround had to be
added to be able to re-add qemu-usermode to MACHINE_FEATURES, since it is removed on the
skylake tune in meta-intel, hence the addition of the qemuuser-enable-chromebook class.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <aehs29@gmail.com>