linux-chromium: Enable building Chromiums Linux kernel fork
Chromium has its own fork of the Linux kernel which they use in
products, add a recipe that allows us to build such kernel at its
latest current version (4.19.70)

The Linux-chromium recipe is only supported on the eve-chromebook
MACHINE at this point.

The config used for this kernel is built from the Intel BSP, adding
a defconfig obtained from the chromiums kernel sources for the
x86_64 flavour (current release is 77, and specific options
obtained from a running Chrome OS kernel on the eve-chromebook.

A vanilla version of these configurations is kept in the
vanilla_kernel_configs directory for documentation purposes.

More information on how the kernel configs are built is kept on
README.kernel.config for the same purpose.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <aehs29@gmail.com>