xkeyboard-config: Modify pixelbook keymap to use right Alt instead of Left Win as third switch
While using the Left Windows key as third switch made more sense to
try and match use the keyboard has in Chrome OS, at the same time
it created some other complications on which the LeftWin key couldnt
be used anymore to invoke the launcher for example.

The Right Alt key may make more sense to use as a third level switch,
and leave the LWIN key for other purposes, including acting as
Caps Lock ironically when the third level switch is used.

Also rename the patch since it adds a keyboard map and not a keyboard.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <aehs29@gmail.com>