Azure Pipelines: Rework CI
Create a matrix of builds depending on the MACHINE and the
kernel used to build an image.

This allows us to have more granularity on the testing
infrastructure making it simpler to find what combinations
are affected by an error.

Create a specific Sstate pipeline to merge all the Shared State
Cache coming from all combinations within the matrix that publishes
a single artifact that can be later consumed by each of the
pipelines in the matrix.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <aehs29@gmail.com>