vboot-utils-chromium: Rework recipe
Build all tools from chromiums vboot_reference repo,
install only the ones we actually need, add a patch
to remove the fuzzer compiler option to allow us
to build all utilities.

Add capability to use as a native recipe which is required
to sign kernels, which would eventually allow us to boot
from u-boot on chromebooks, this is a requirement for
ARM based chromebooks.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <aehs29@gmail.com>