linux-chromium: Create recipe for linux-chromium 3.18
This recipe builds Chromiums kernel fork 3.18, it is currently
only supported on the mediatek8173-chromebook MACHINE.

A defconfig is used, which was taken from Arch Linux oak kernel package,
the reason this defconfig is used is simply because it boots properly,
a copy of this defconfig is also added on the vanilla_kernel_configs
directory.

IMO a proper config should be generated the same way the configs for other
chromebooks are generated, although I have tried to use the same process
using the mediatek flavour, but was unsuccessful on creating a bootable
kernel config neither on the stock oak branch nor the R80 latest branch,
I also got a kernel config from a running Chrome OS R79 on the
Chromebook itself, but was unable to boot using that one either.

Since I only had the Chromebook available for some time, I was forced to
keep Archs oak config instead of debugging the problem with the generated
configs, perhaps in the future if I have access to a similar device I will
be able to debug and generate the config properly.

A copy of the generated configs and the running config is kept on the
vanilla_kernel_configs directory for reference.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <aehs29@gmail.com>