Seabios: Get date only once
Previously the date command was executed several times
within the recipe, the build system was expecting the
same result everytime, but, while it might seem uncommon
the result was not always the same depending on when the
shell was launched, causing errors like the following:

install: cannot stat SeaBIOS-kbl_20200101.bin*: No such
file or directory.

Get the date only once during parsing and keep that variable
across tasks to fix the issue.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <aehs29@gmail.com>