seabios: Fix build for binutils 2.36
When upgrading to binutils 2.36, two errors are thrown out:

| x86_64-poky-linux-ld: cannot use executable file out/rom16.strip.o as input to a link
| make: *** [Makefile:190: out/rom.o] Error 1

and

x86_64-poky-linux-ld: section .note.gnu.property LMA [0000000000000000,0000000000000027]
overlaps section .text LMA [0000000000000000,0000000000006d7c]
| make: *** [Makefile:242: out/vgarom.o] Error 1

THe first one needs a patch backported from upstream, while the second one
either needs to be fixed in the linker script vgalayout or the section
gnu.property needs to be removed from the object files to avoid collisions.

This patch uses objcopy to remove the sections since it is less disruptive,
will wait until this is fixed upstream to backport a fix for the linker
script.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandro@enedino.org>